Simplify the logging classes considerably, and tweak the logging macros to skip the output stream formating for messages that aren't logged (due to low priority or category filtering). Previously, all messages were formatted and sent either to the real log, or dropped (one character at a time) by the underlying streambuf.
When debugging is disabled, the log statements now reduce to a
form that produces no code but should prevent spurious compiler
warnings about unused variables (for local variables that are
only referenced in the log output).

Changed the makefile to always rebuild Version.cpp, so the
'build on [date]' message will be accurate.

git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1152 7152ba15-5bd8-0310-b828-e896570a432b