* More gear animation refactoring.  Merged the gear sequence and structure animations into a single class.  Split part of TimedSequence into a SharedSequence base class and created a DrivenSequence subclass.
* Added rate limiting for steering.  This is just a temporary measure until
steering control is intergrated into the fcs.

* Adjusted gear stiffness, damping, and position for the F16.

* Set the 3D model offset to zero for both the M2k and F16 and added a
center of mass offset to DynamicObject.  GlobalPosition now represents
the model position and is exported as the Kinetics::ModelPosition channel.
The center of mass position replaces the Kinetics::Position channel and is
used for physics calculations.  An extra transform node at the top of the
scene model provides the necessary offset to rotate around the center of
mass.

* Added a new type of debug marker ("debug_points") to ObjectModel.  These
are similar in appearance to contact points (actually, small yellow diamonds
instead of large red diamonds) but have no physical interaction.  Should be
useful for testing landing gear position and view point.

===> ALL USERS: RebuildData.py

git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1521 7152ba15-5bd8-0310-b828-e896570a432b