* Add support for serializing uninitialized TableN instances, which allows TableN members to be marked as optional in an XML interface.
* Add better support for TableN instances in the XML parser.  The parser
now checks that the number of values matches the count expected from
the breakpoints.  Breakpoints and table values can now be scaled by
the parser (e.g., scale="0.5" in the BreakN and Values attributes),
which for example allows radian angles to be conveniently specified in
degrees in the XML file.  Note that the "spacing" attribute for
breakpoints is also scaled by this factor.

* Fix a syntax error in Link.h.

git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1624 7152ba15-5bd8-0310-b828-e896570a432b