* Refactor existing flight model code and add a new flight model implementation for the f-16 based on NASA low speed wind tunnel data and Raptor/Mav-JP's high speed flight model data.  The FlightModel class is now abstract, with StabilityFlightModel implementing a concrete subclass that is similar to the old FlightModel class (mostly linear stability derivatives).  The F16FlightModel subclass implements more specialized routines based on the NASA 1979 paper, using tabular data for most coefficients over a broad range of AOA (-20 to +90 deg).  This includes deep stall modeling.
Improved modeling for the f-16 fcs yaw limiter.  Also improved
the pitch limiter in TO/landing gains.

Changed the control surface deflection sense to match the
literature.  Updated the model animations accordingly.

Updated animation gains to compensate for an earlier change to the
deflection, limit, and rate scaling.  Rotations are now assumed
to be in degrees, so the control surface deflections need to be
prescaled by 180/PI.

Added the MPO switch and made minor realism tweaks to the f-16 flcs.

This change breaks the air refueling door.  That logic is moving to
the new fuel system code which will be submitted in a later change.

git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1643 7152ba15-5bd8-0310-b828-e896570a432b