Refactor shared libraries.  The csplib and cspsim libraries are now built separately from their python wrapper modules.  Runtime library resolution is now done via rpath under Linux, rather than RTLD_GLOBAL and ordered python imports.  This will hopefully fix import errors on Linux platforms that don't support RTLD_GLOBAL from python.  Library resolution under Windows still uses PATH overrides.
Incorporated the layout tool into the main build system.  Run 'scons -u
layout' to build the layout tool (it is not part of the 'all' target yet).

Added various *_EXPORT attributes to fix symbol exports under Windows.

Changed InterfaceRegistry to not derive from Singleton<>.  Singleton<>
requires that the instance variable be declared in a header, which can
cause problems with duplicate initialization across multiple modules
that use csplib.  Similarly, constructors for two theater classes had
to be moved to .cpp files to avoid dynamic_cast failures when used in
the layout tool, and several template classes in Random.h had to be
explicitly instantiated to export correctly under Windows.

git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1955 7152ba15-5bd8-0310-b828-e896570a432b