Hook in new sky and shader code.  This will disable per-pixel specular highlights on graphics cards that don't support glsl, but otherwise shouldn't cause problems.
On cards that support shaders, all objects are rendered in
shaders.  Default shaders are automatically bound based on
the model statesets when possible.  Currently, all geodes
must have a Material attribute, and may also have a 2D or
cubemap texture bound to texture unit 0.  Geodes that do
not fall into these categories are rendered in red to help
identify them.

Empty Program {} nodes can be attached to geodes to force
fixed function rendering, but note that this bypasses global
fog rendering.  This is done for light points and near-field
models like the detailed cockpit and mfd/hud displays.

The terrain drawables need special shaders to compute per-
pixel fog on the gpu. (For normal objects, fog is a uniform
computed on the cpu.)  This has been done for demeter, but
not for chunklod (yet).

Included in this change is a minor adjustment to bright-star
rendering to rescale the flare imposters so that their size
is independent of the viewing angle.

Added VASI lights and models to the airbase.

Remodelled the light towers in blender with extensive hand
editing of the exported osg files to make them more efficient.
Added blinking red lights at the top and steady white lights.

Adjusted the sim start time to pre-dawn.

git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@1982 7152ba15-5bd8-0310-b828-e896570a432b