Merged new_look_and_feel branch into the trunk. The following revisions were taken r2109:2121.
* Hover effects on buttons. When the mouse is positioned over a button it will be displayed with an other style.
* A completely new theme with background images for buttons and the desktop. 
Icons has been copied from the free Tango project. Original svg files is also added to subversion if we want to rescale the images.

The main menu is still located to the left with the following options:
* Instant action
* Tutorials (disabled button)
* Single player (disabled button)
* Multiplayer (disabled button)
* Credits (disabled button)

A new toolbar has been added that is aligned to the top right corner. The following icons has been added.
* Help
* Library (disabled button)
* Options
* Quit

The help button opens a window with information on how to take off and what the key bindings are. All texts has been copied from our wiki site.

All user interface windows has been adjusted to look best in 1024x768 and higher resolutions. If a lower resolution is used it still works but has lower quality. Some lines that should be 1 pixel in width will be less and disapears. Texts will be less readable etc.


git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@2122 7152ba15-5bd8-0310-b828-e896570a432b