Solves the scons build issue with: "TypeError: unsupported operand type(s) for +: 'Executor' and 'unicode'".
We create two phony build targets (config, default). And we tried to set
a message to be displayed when any of those target was invoked. For some
reason scons fails to handle the display of message. Might be a bug or
changed behaviour from scons.

The config target is now an empty list (that won't output anything).

The default target is removed completely (it printed "No targets
specified; try scons -h for help."). Now scons will try to build
"all" target if no arguments are specified.