Fix font crash, shaders and demeter linking error.
It is mandatory to call addGlyph from one of the newer OpenSceneGraph
versions. If not you will get a crash when retreiving the texture for
a glyph (is assigned in addGlyph).

Removed 'const' from various shaders since we got a lot of errors
from them when running the simulator.

Added libGLU to the demeter module. This is needed in order for it
to generate the terrain.