Removed the old csp bootstrap module.
The 'csp' python module should always be findable by setting
the PYTHONPATH member variable.

The old csp bootstrap module was installed globally and did
some clever trix to replace itself based on the actuall directory
you was in. The only purpose was so you didn't have to set
the PYTHONPATH variable or change it if you had multiple copies
of csp on your file system.