Use boost::asio for networking
This commit rewrites all networking code to use boost::asio instead
of the older Gnu Common C++ library. This also drops the dependency
to that library.

Starting the IndexServer and connect two clients will create network
traffic between the server and clients. Clients will be notified with
a message that someone has connected. Occasionally you can see the
other client but not always. This is the state before this refactor
so this hasn't made things worse at least.

Further bugfixing is needed in order for the IndexServer to work
properly.