Fix so only one TaggedRecordRegistry was created
This is a singleton class and only one should be created. Since the
singleton template only exists in the header file this resulted in two
instances. One for csplib and one for cspsim. This in turn caused
network messages that cspsim created to not be recognized when
running.

Using a separate function for retrieving the singleton instance and
make sure that it is placed in a .cpp file will force a single
instance. Also preventing access to the original getInstance method by
doing a private inheritance.